[BUGFIX] Switch to List module on root page leads to blank page
Reproducable only with admin account. Switch to Page module, go to root
page (id = 0), switch to List module => blank page. Works only on
root page.

Releases: 6.0, 4.7, 4.6, 4.5
Resolves: #40781
Reviewed-on: http://review.typo3.org/14799
Tested-by: Markus Klein
Tested-by: Georg Ringer

+++ typo3/sysext/recordlist/mod1/index.php
@@ -128,7 +128,7 @@ class SC_db_list {
 		$this->perms_clause = $BE_USER->getPagePermsClause(1);
 
 			// GPvars:
-		$this->id = t3lib_div::_GP('id');
+		$this->id = (int) t3lib_div::_GP('id');
 		$this->pointer = t3lib_div::_GP('pointer');
 		$this->imagemode = t3lib_div::_GP('imagemode');
 		$this->table = t3lib_div::_GP('table');
